% Generated by roxygen2 (4.0.0): do not edit by hand
\name{plot_clade}
\alias{plot_clade}
\title{Plot frequency of response for a single clade}
\usage{
plot_clade(peptideSet, calls, clade, from = 0, to = max(end(peptideSet)))
}
\arguments{
\item{peptideSet}{A \code{peptideSet} object.}

\item{calls}{A \code{matrix}, as returned by the \code{makeCalls} function.}

\item{clade}{A \code{character}, the name of the clade to plot. It must be a
valid column name of \code{clade(peptideSet)}.}

\item{from}{A \code{numeric}, the start coordinate of the plot.}

\item{to}{A \code{numeric}, the end coordinate of the plot.}
}
\description{
Plot an axis and the frequency of response of a single selected clade
}
\details{
The peptideSet should be the one used in the function call to \code{makeCalls}
that generated the calls used. They should have identical peptides in the same
order.
The function requires the calls matrix to have been created using \code{makeCalls}
with freq set to TRUE and a grouping variable.
}
\author{
Renan Sauteraud
}
\seealso{
plot_inter
}

