% Generated by roxygen2 (4.0.0): do not edit by hand
\name{plot_inter}
\alias{plot_inter}
\title{Plot frequency of response for each group}
\usage{
plot_inter(peptideSet, calls, from = 0, to)
}
\arguments{
\item{peptideSet}{A \code{peptideSet} object.}

\item{calls}{A \code{matrix}, as returned by the \code{makeCalls} function.}

\item{from}{A \code{numeric}, the start coordinate of the plot.}

\item{to}{A \code{numeric}, the end coordinate of the plot.}
}
\description{
Plot an axis and the frequency of response of each group, averaged by peptides
at each position.
}
\details{
The peptideSet should be the one used in the function call to \code{makeCalls}
that generated the calls used. They should have identical peptides in the same
order.
The function requires the calls matrix to have been created using \code{makeCalls}
with freq set to TRUE and a grouping variable.
}
\author{
Renan Sauteraud
}
\seealso{
plot_clade
}

