% Generated by roxygen2 (4.0.0): do not edit by hand
\name{shinyPepStat}
\alias{shinyPepStat}
\title{Launch the pepStat Shiny Application}
\usage{
shinyPepStat()
}
\description{
Launches the \code{pepStat} Shiny application, providing an interactive
interface for constructing peptide sets, normalizing intensities, generating
calls. Quality control is also facilitated through interactive plotting
features.
}
\examples{
if (interactive()) {
  shinyPepStat()
}
}

