% Generated by roxygen2 (4.0.0): do not edit by hand
\name{baselineCorrect.pSet}
\alias{baselineCorrect.pSet}
\title{Substract baseline intensities}
\usage{
baselineCorrect.pSet(pSet, verbose = FALSE)
}
\arguments{
\item{pSet}{A \code{peptideSet} with sample PRE and POST visits.}

\item{verbose}{A \code{logical}. If TRUE, information regarding the
pairedness of the data will be displayed.}
}
\value{
A \code{matrix} of the baseline corrected intensities, with as many
columns as there are samples POST visit
}
\description{
Correct intensities by substracting PRE visit sample intensities.
}
\details{
If samples are not PAIRED (One PRE and POST for each ptid), then the average
expression of all PRE visit samples is substracted from each sample.
}
\author{
Raphael Gottardo, Gregory Imholte
}

