% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{peptideSet}
\alias{peptideSet}
\alias{peptideSet-class}
\title{peptideSet class}
\description{
This class gathers all information from gpr files, annotation data and sequence data
}
\details{
See \code{?`peptideSet-methods`} for a list of accessors and method associated
with the class.
}
\section{Slots}{

\describe{
 \item{featureRange}{A \code{GRanges}object. The ranges and sequences of
the peptides and their associated annotation.}
 \item{phenoData}{An \code{AnnotatedDataFrame}. Annotation for the samples.}
 \item{assayData}{}
 \item{featureData}{}
 \item{annotation}{}
 \item{protocolData}{Slots inherited from \code{ExpressionSet}.}
}
}

\seealso{
\code{\link{ExpressionSet}}, \code{\link{peptideSet-methods}}
}
\author{
Greg Imholte
}
