% Generated by roxygen2 (4.0.0): do not edit by hand
\name{restab}
\alias{restab}
\title{Result table}
\usage{
restab(peptideSet, calls, long = FALSE)
}
\arguments{
\item{peptideSet}{A \code{peptideSet} object.}

\item{calls}{A \code{matrix}, as returned by the \code{makeCalls} function.}

\item{long}{A \code{boolean}. If set to TRUE, the result table will have one
row per peptide/clade. If FALSE, one row per peptide and all clades are
listed together.}
}
\value{
A \code{data.frame} with the peptides and some information from the
\code{peptideSet} as well as the frequency of binding for each group of the
calls.
}
\description{
Tabulate the results of a peptide microarray analysis.
}
\details{
The peptideSet should be the one used in the function call to \code{makeCalls}
that generated the calls used. They should have identical peptides.
}

