\name{slidingMean}
\alias{slidingMean}

\title{Data smoothing for peptide microarrays}

\description{
  This function uses a sliding window correct the noise generated by the experimental variation. It take advantage of the overlapping nature of the peptide on the arrayi to reduce signal variability. Two overlapping peptides binding effexts are expected to be positively correlated.
}

\usage{
slidingMean(peptideSet, width=5, verbose=FALSE)
}

\arguments{
  \item{peptideSet}{A \code{peptideSet} object. The data.}
  \item{width}{A \code{numeric}. The width of the sliding window.}
  \item{verbose}{A \code{logical}. If set to TRUE, progress information will be displayed.}
}

\value{
The function returns the peptideSet object with modified intensities. The distribution of the new intensities is centered around zero.
}


\seealso{
\code{NormalizeArray}
}

\keyword{functions}
