\name{slidingMean}
\alias{slidingMean}
\title{Data smoothing for peptide microarray.}
\usage{
slidingMean(peptideSet, width = 5, verbose = FALSE, split.by.space = TRUE)
}
\arguments{
  \item{peptideSet}{A \code{peptideSet}. The expression
  data for the peptides as well as annotations and ranges.
  The range information is required to run this function.}

  \item{width}{A \code{numeric}. The width of the sliding
  window.}

  \item{verbose}{A \code{logical}. If set to TRUE, progress
  information will be displayed.}

  \item{split.by.space}{A \code{logical}. If TRUE, the
  peptides will be smoothed by space (See details section
  below and \code{RangedData} for more information).}
}
\value{
A \code{peptideSet} object with smoothed intensities.
}
\description{
This function applies a sliding mean window to intensities
to reduce noise generated by experimental variation, as
well as take advantage of the overlapping nature of array
peptides to share signal.
}
\details{
Peptide membership in the sliding mean window is determined
by its position and the width argument. Two peptides are in
the same window if the difference in their positions is
less than or equal to width/2. A peptide's position is
taken to be position(peptideSet).

A peptide's intensity is replaced by the mean of all
peptide intensities within the peptide's sliding mean
window.

When split.by.space = TRUE, peptides are smoothed within
groups defined by the space of the RangedData object
occupying the featureRange slot of peptideSet. This is
useful if, for example, spaces indicate peptides drawn from
separate proteins that should not be smoothed together.
}
\author{
Gregory Imholte
}
\seealso{
\code{\link{summarizePeptides}},
\code{\link{normalizeArray}}
}

